requirements.txt
flask==3.1.2
langchain==1.2.8
langchain-community==0.4.1
langchain-core==1.2.9
langchain-openai==1.1.7
langchain-pinecone==0.2.13
pinecone==7.3.0
pypdf==6.6.2
python-dotenv==1.0.1
sentence-transformers==5.2.2
-e .